refactor standupservice
